requirements.txt
Django>=1.11, <2
django-js-reverse>=0.8.1
django-webpack-loader>=0.6.0
djangorestframework>=3.7.7, <4
unicode-slugify>=0.1.3
whitenoise>=3.3.1
